feat(HTTP Server): implemented a HTTP server
This is the first time stuff has worked while using Rust and Cargo.
`wasm-pack` was just too weird and cumbersome to use. `tiny_http` has
proven useful and we'll proceed to spawning a Web Socket Server(WSS) using
JavaScript.

The WSS will be the main mode of how we'll be communicating with Deriv's
API. Rust will be the framework by which our Application will run.

TODO: Implement a simple WSS
Assigned-to: @m-ikiara
Depends-on: #6b42ad7